Semver Utility for DB storage
Working with versions of various objects is a common problem, which already has
a number of market-adopted solutions. One of these solutions is Semantic
Versioning - a system of rules and requirements for assigning version numbers
to software components and other objects. One can find the specification for
SemVer freely at http://semver.org

It is proposed to add support of the semantic versioning concept into Glance,
so Glance objects (starting from Artifacts, but probably including Images in
future) may be properly versioned.

Change-Id: I72f7233f09ea36556fd46cab389988d1698eac37