Remove `target_*` fields
A cloud must choose what the best target format for its own deployment
is without requiring the user to pass this information over. A user
shouldn't worry about what the target formats in a cloud are. To some
extent, the user should not care about the image being converted at all.

This patch removes the requirement of passing `target_` fields in the
import call but it keeps them, as a way to inform the user how the cloud
operates, in the `info` call.

Glance already provides a way for cloud operators to specify what format
images should be converted to: `conversion_format`

MitakaPriority
Change-Id: I60e2987e430a2f4a2e0dc3f34e96577a1c854364