Remove reviewers section from template
When this template was created, the intention behind this section was to
know the got-to people for specific specs. Questions like: Who should I
ping if I need more Glance insights? Who should I ping to actually get
reviews? etc

Unfortunately, The reviewers section has turned to a field mostly used
to highlight people interested in a specific feature. More often than
not, the revieweres listed in this section are not the ones reviewing
the patches for the spec.

In addition to the above, this section currently separates core
reviewers from non-cores. This separation brings no value to this
section nor the spec itself.

Regardless the size of the team, this section is not bringing any value
to the process as reviewers not in this list can as well review the
patches as much as the reviewers listed there can also not review the
patches.

Change-Id: I973d97e82f4561de8d4ac33672494eae0fc15925