Return 409 if removing/replacing image locations.
Currently image location updates (removing, replacing) are permitted for
images even if their state is not ``active``.
This lite-spec describes the need to prevent removal of image
locations of images that are not in ``active`` state and the replacement of
image locations of images that are not in ``active`` or ``queued`` state by
returning a Conflict Error (409 response code).

Change-Id: I6d984fcafa49ad806dce2bd56abcee66fdc77525