Pending-delete rollback ability
Glance support soft delete images with delayed_delete feature. If
this feature is open, when users delete a image, the image will
first be in a special 'pending_delete' status that is not displayed
in the API response. And then the image will be deleted by
glance-scrubber process.

But currently, there is no way to revert the pending-delete images.
Once the admin operators want to revert the delete action, Glance
should give them this kind of ability.

This patch will enhance the glance-scrubber tool to support restoring
the image from pending-delete to active. Although, whether the rollback
of a particular image is possible or not depends upon Glance configuration
option settings and quick operator action.

Change-Id: I0202b754815a3d2133a57ef987741c7b12c29ff3
blueprint: pending-delete-rollback