Open specs repo for Rocky
Glance is now in a specs proposal freeze for Queens, so new specs
must be proposed for Rocky.  This patch sets up the directories
and introduces a new spec-lite template to replace the old strategy
of having a single file containing all lite specs.  It also reorganizes
the index page slightly to make the current cycle's specs more
prominent.

Change-Id: I85b096fc0c99a70e3ad0d7f8727b7d5ab679c95b