Modify operator image workflow specs
Modified specs to reflect changes as per the implementation.
Highlights:
In the response of create/show image the new property will be
displayed as 'os_hidden'.
If a pre-Rocky image already has a custom property named as
'os_hidden' then that property will no longer be visible in
the response from Rocky release.

Change-Id: I9d012a81ad5c8d209cf52634740268989eaed928