Remove setup.py and setup.cfg
Theses used to be used for building docs in the past.

A recent release of setuptools 61.0.0 introduced a breaking change [1]
which caused our openstack-tox-docs and openstack-tox-pep8 jobs to fail
with:

  error: Multiple top-level packages discovered in a flat-layout:
    ['specs', 'priorities'].

This removes setup.py and setup.cfg from use in tox.ini to resolve the
issue and they are not needed anyway.

Related-Bug: #1966382
Change-Id: Ia653b3e13302db5bf2adc006aea5436a9c65616e